fastlane-finale-service/pyproject.toml
[tool.poetry]
name = "fastlane-finale-service"
version = "0.1.0"
description = ""
authors = ["Patrick Kalkman <patrick@simpletechture.nl>"]
readme = "README.md"
packages = [{include = "fastlane_finale_service"}]

[tool.poetry.dependencies]
python = "^3.11"
loguru = "^0.7.0"
paho-mqtt = "^1.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
